ITS-3010 Removed ConfigDecorator
